Fix editor review items and other cleanup
Simplifies the code a bit for presentation in doc content.
