+++ swift/example_code/s3/ListBuckets/Package.swift
@@ -8,11 +8,14 @@
 import PackageDescription
 
 let package = Package(
+    // snippet-start:[s3.swift.listbuckets.name-platforms]
     name: "listbuckets",
     platforms: [
         .macOS(.v11),
         .iOS(.v13)
     ],
+    // snippet-end:[s3.swift.listbuckets.name-platforms]
+    // snippet-start:[s3.swift.listbuckets.dependencies]
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         .package(
@@ -22,32 +25,33 @@ let package = Package(
         .package(
             url: "https://github.com/apple/swift-argument-parser.git",
             .branch("main")
-        ),
-        .package(
-            name: "SwiftUtilities",
-            path: "../../../modules/SwiftUtilities"
-        ),
+        )
     ],
+    // snippet-end:[s3.swift.listbuckets.dependencies]
+    // snippet-start:[s3.swift.listbuckets.targets]
     targets: [
         // A target defines a module or a test suite. A target can depend on
         // other targets in this package. They can also depend on products in
         // other packages that this package depends on.
+        // snippet-start:[s3.swift.listbuckets.targets-executable]
         .executableTarget(
             name: "listbuckets",
             dependencies: [
                 .product(name: "AWSS3", package: "aws-sdk-swift"),
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
             ],
-            path: "./Sources"
+            path: "./Sources/ListBuckets"
         ),
+        // snippet-start:[s3.swift.listbuckets.targets-test]
         .testTarget(
             name: "listbuckets-tests",
             dependencies: [
                 .product(name: "AWSS3", package: "aws-sdk-swift"),
-                "listbuckets",
-                "SwiftUtilities"
+                "listbuckets"
             ],
-            path: "./Tests"
+            path: "./Tests/ListBucketsTests"
         )
+        // snippet-end:[s3.swift.listbuckets.targets-test]
     ]
+    // snippet-end:[s3.swift.listbuckets.targets]
 )
